[parquet-sdk]migrate objects [pt1] (#640)
