+++ rust/processor/src/db/common/models/fungible_asset_models/raw_v2_fungible_metadata.rs
@@ -7,11 +7,13 @@
 
 use crate::{
     db::{
-        common::models::token_v2_models::v2_token_utils::TokenStandard,
+        common::models::{
+            object_models::v2_object_utils::ObjectAggregatedDataMapping,
+            token_v2_models::v2_token_utils::TokenStandard,
+        },
         postgres::models::{
             coin_models::coin_utils::{CoinInfoType, CoinResource},
             fungible_asset_models::v2_fungible_asset_utils::FungibleAssetMetadata,
-            object_models::v2_object_utils::ObjectAggregatedDataMapping,
             resources::FromWriteResource,
         },
     },
