+++ rust/processor/src/db/common/models/mod.rs
@@ -3,4 +3,5 @@ pub mod ans_models;
 pub mod default_models;
 pub mod event_models;
 pub mod fungible_asset_models;
+pub mod object_models;
 pub mod token_v2_models;
