+++ rust/processor/src/db/common/models/object_models/mod.rs
@@ -1,4 +1,5 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod v2_objects;
+pub mod raw_v2_objects;
+pub mod v2_object_utils;
