+++ rust/processor/src/db/common/models/object_models/v2_object_utils.rs
