+++ rust/processor/src/db/common/models/token_v2_models/raw_v2_token_datas.rs
@@ -7,9 +7,11 @@
 
 use crate::{
     db::{
-        common::models::token_v2_models::v2_token_utils::{TokenStandard, TokenV2, TokenV2Burned},
-        postgres::models::{
+        common::models::{
             object_models::v2_object_utils::ObjectAggregatedDataMapping,
+            token_v2_models::v2_token_utils::{TokenStandard, TokenV2, TokenV2Burned},
+        },
+        postgres::models::{
             resources::FromWriteResource, token_models::token_utils::TokenWriteSet,
         },
     },
