+++ rust/processor/src/db/postgres/models/resources.rs
@@ -2,9 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::db::{
-    common::models::token_v2_models::v2_token_utils::{
-        AptosCollection, Collection, ConcurrentSupply, FixedSupply, PropertyMapModel,
-        TokenIdentifiers, TokenV2, UnlimitedSupply,
+    common::models::{
+        object_models::v2_object_utils::{ObjectCore, Untransferable},
+        token_v2_models::v2_token_utils::{
+            AptosCollection, Collection, ConcurrentSupply, FixedSupply, PropertyMapModel,
+            TokenIdentifiers, TokenV2, UnlimitedSupply,
+        },
     },
     postgres::models::{
         default_models::move_resources::MoveResource,
@@ -12,7 +15,6 @@ use crate::db::{
             ConcurrentFungibleAssetBalance, ConcurrentFungibleAssetSupply, FungibleAssetMetadata,
             FungibleAssetStore, FungibleAssetSupply,
         },
-        object_models::v2_object_utils::{ObjectCore, Untransferable},
     },
 };
 use anyhow::Result;
